Conflicts:
	contest/views.py
修复冲突，修改了不恰当的函数名

+++ tools/release_static.py
@@ -9,6 +9,12 @@ template_release_path = "template/release/"
 
 static_src_path = "static/src/"
 static_release_path = "static/release/"
+
+print "Begin to compress js"
+if os.system("node static/src/js/r.js -o static/src/js/build.js"):
+    print "Failed to compress js, exit"
+    exit()
+
 try:
     # 删除模板的 release 文件夹
     shutil.rmtree(template_release_path)
@@ -30,7 +36,6 @@ name_map = {}
 
 def process(match):
     file_path = match.group(1).replace("/static/", "")
-    # print file_path, match.group(), match.group(1)
 
     if not os.path.exists(static_release_path + file_path):
         return match.group(0)
@@ -49,9 +54,11 @@ def process(match):
         return match.group(0)
 
 
+print "Begin to add md5 stamp in html"
 for root, dirs, files in os.walk(template_release_path):
     for name in files:
         html_path = os.path.join(root, name)
+        print "Processing: " + html_path
         html_content = open(html_path, "r").read()
         js_replaced_html_content = re.sub(js_re, process, html_content)
         css_replaced_html_content = re.sub(css_re, process, js_replaced_html_content)
@@ -59,3 +66,5 @@ for root, dirs, files in os.walk(template_release_path):
         f = open(html_path, "w")
         f.write(css_replaced_html_content)
         f.close()
+
+print "Done"
